+++ tests/bot/presenceRouting.test.ts
@@ -196,6 +196,7 @@ describe("presence routing", () => {
     ["restart", {}],
     ["dev_raid_stop", {}],
     ["dev_reset_monster_rest", {}],
+    ["dev_add_bandage", {}],
     ["dev_reset_yeger_bandage", {}],
     ["dev_help", {}]
   ])("routes command /%s", (command, expected) => {